feat: cleanup code background image
