+++ apps/codeimage/src/components/PropertyEditor/controls/CustomColorPicker.css.ts
@@ -1,10 +1,4 @@
-import {
-  backgroundColorVar,
-  textFieldStyles,
-  themeVars,
-  withThemeMode,
-} from '@codeimage/ui';
-import {darkGrayScale} from '@codeimage/ui/themes/darkTheme';
+import {backgroundColorVar, textFieldStyles, themeVars} from '@codeimage/ui';
 import {style} from '@vanilla-extract/css';
 import {recipe, RecipeVariants} from '@vanilla-extract/recipes';
 
@@ -58,81 +52,6 @@ export const colorItem = recipe({
 export const popover = style({
   maxWidth: 'initial',
   transition: `all 250ms ease-in-out`,
-});
-
-export const imageGrid = style([
-  {
-    display: 'grid',
-    padding: themeVars.spacing['2'],
-    gridTemplateColumns: 'repeat(1, 1fr)',
-    gap: themeVars.spacing[3],
-    overflow: 'auto',
-    maxHeight: '400px',
-    width: '240px',
-    '@media': {
-      '(min-width: 640px)': {
-        width: '544px',
-        gridTemplateColumns: 'repeat(3, 1fr)',
-      },
-    },
-  },
-]);
-
-export const imagesCard = style([
-  {
-    backgroundColor: themeVars.dynamicColors.input.backgroundColor,
-    width: '100%',
-    borderRadius: themeVars.borderRadius.xl,
-    padding: themeVars.spacing['5'],
-    paddingBottom: themeVars.spacing['2'],
-    boxShadow: themeVars.boxShadow.md,
-    color: themeVars.dynamicColors.baseText,
-    transition: 'background-color 0.2s ease-in-out',
-    position: 'relative',
-    display: 'inline-flex',
-    flexDirection: 'column',
-    selectors: {
-      '&[data-active]': {
-        outline: `2px solid ${themeVars.dynamicColors.primary}`,
-        outlineOffset: '2px',
-      },
-      ...withThemeMode({
-        dark: {
-          backgroundColor: themeVars.dynamicColors.input.backgroundColor,
-          border: `1px solid ${darkGrayScale.gray7}`,
-        },
-        light: {
-          backgroundColor: themeVars.backgroundColor.white,
-          border: `1px solid ${themeVars.borderColor.default}`,
-          boxShadow: themeVars.boxShadow.default,
-        },
-      }),
-
-      ...withThemeMode(
-        {
-          dark: {backgroundColor: darkGrayScale.gray2},
-          light: {backgroundColor: darkGrayScale.gray12},
-        },
-        '&:hover',
-      ),
-    },
-  },
-]);
-
-export const textFieldImageUrl = style({
-  borderTopRightRadius: 0,
-  borderBottomRightRadius: 0,
-  flex: 1,
-});
-
-export const textFieldImageInput = style({
-  borderTopRightRadius: 0,
-  borderBottomRightRadius: 0,
-});
-
-export const confirmButton = style({
-  borderTopLeftRadius: 0,
-  borderBottomLeftRadius: 0,
 });
 
 export type ColorPickerColorItemProps = RecipeVariants<typeof colorItem>;
